refactoring yolo trainer
