+++ configs/pipelines/train_detector_mit_yolo_640.conf
@@ -1,0 +1,105 @@
+#  Groundtruth file extensions (txt, kw18, etc...). Note: this is indepedent of
+#  the format that's stored in the file.
+groundtruth_extensions = .csv
+
+#  Algorithm to use for 'groundtruth_reader'.
+#  Must be one of the following options:
+#  	- habcam :: Reads habcam detection/ground truth files.
+#  	- kw18 :: Detected object set reader using kw18 format.
+#  	- viame_csv :: Detected object set reader using VIAME csv format.
+groundtruth_reader:type = viame_csv
+
+#  Dump possible input data formatting warnings to these files
+data_warning_file = TRAINING_DATA_WARNINGS.txt
+groundtruth_reader:viame_csv:warning_file = TRAINING_DATA_WARNINGS.txt
+
+#  Can be either: "one_per_file" or "one_per_folder".
+groundtruth_style = one_per_folder
+
+#  Semicolon list of seperated image extensions to use in training, images
+#  without this extension will not be included.
+image_extensions = .jpg;.jpeg;.JPG;.JPEG;.tif;.tiff;.TIF;.TIFF;.png;.PNG;.sgi;.SGI;.bmp;.BMP;.pgm;.PGM
+
+#  Semicolon list of seperated video extensions to use in training, videos
+#  without this extension will not be included.
+video_extensions = .mp4;.MP4;.mpg;.MPG;.mpeg;.MPEG;.avi;.AVI;.wmv;.WMV;.mov;.MOV;.webm;.WEBM;.ogg;.OGG
+
+#  Pipeline to use to extract video frames if inputs are videos
+relativepath video_extractor = filter_default.pipe
+
+#  Percent [0.0, 1.0] of validation samples to use if no manual files specified.
+default_percent_validation = 0.30
+
+#  Number of validation frames to group together in one test burst
+# validation_burst_frame_count = 10
+
+#  Augmentation pipeline.
+#relativepath augmentation_pipeline = train_hue_shifting_only_aug.pipe
+
+#  Augmentatation cache.
+augmentation_cache = augmented_images
+
+#  Do not regenerate cache.
+regenerate_cache = true
+
+
+#  Algorithm to use for 'detector_trainer'.
+detector_trainer:type = ocv_windowed
+
+block detector_trainer:ocv_windowed
+
+  # Directory for all files used in training
+  train_directory = deep_training
+
+  # Windowing mode, can be disabled, maintain_ar, scale, chip, adaptive
+  mode = disabled
+
+  # Resized chip width.
+  chip_width = 640
+
+  # Resized chip height.
+  chip_height = 640
+
+  # Adaptive size threshold for chipping
+  chip_adaptive_thresh = 1600000
+
+  # Don't train on chips with detections smaller than this
+  min_train_box_length = 5
+
+  # Uncomment to remove small detections instead of training on them
+  small_box_area = 200
+  small_action = remove
+
+  # Image reader type
+  image_reader:type = vxl
+
+endblock
+
+block detector_trainer:ocv_windowed:trainer
+
+  # Trainer type
+  type = mit_yolo
+
+  # The architecture
+  mit_yolo:model = v9-c
+
+  # Number of GPUs to use, -1 indicates all
+  mit_yolo:gpu_count = -1
+
+  # Network perspective field size
+  mit_yolo:chip_width = 640
+
+  # Path to pretrained state
+  relativepath mit_yolo:weight = true
+
+  #  Pipeline template file.
+  relativepath mit_yolo:pipeline_template = templates/embedded_netharn.pipe
+
+  # Image train batch size
+  mit_yolo:batch_size = 4
+
+  # Training learning rate
+  netharn:learning_rate = 1e-3
+
+endblock
+
